refactor iceberg sink test
