+++ java/connector-node/risingwave-connector-test/pom.xml
@@ -14,6 +14,7 @@
 
     <properties>
         <testcontainers.version>1.17.6</testcontainers.version>
+        <iceberg.version>1.0.0</iceberg.version>
     </properties>
 
     <dependencies>
@@ -25,6 +26,11 @@
         <dependency>
             <groupId>com.risingwave.java</groupId>
             <artifactId>risingwave-sink-deltalake</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.risingwave.java</groupId>
+            <artifactId>risingwave-sink-iceberg</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -69,6 +75,18 @@
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-sql_2.12</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.iceberg</groupId>
+            <artifactId>iceberg-spark-3.2_2.12</artifactId>
+            <version>${iceberg.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.iceberg</groupId>
+            <artifactId>iceberg-arrow</artifactId>
+            <version>${iceberg.version}</version>
             <scope>test</scope>
         </dependency>
         <dependency>
