+++ java/connector-node/risingwave-connector-test/src/test/java/com/risingwave/connector/sink/iceberg/IcebergSinkLocalTest.java
@@ -12,30 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.risingwave.connector;
+package com.risingwave.connector.sink.iceberg;
 
 import static com.risingwave.proto.Data.*;
 import static org.junit.Assert.*;
 import static org.junit.Assert.assertEquals;
 
 import com.google.common.collect.Iterators;
-import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import com.risingwave.connector.IcebergSink;
 import com.risingwave.connector.api.TableSchema;
 import com.risingwave.connector.api.sink.ArraySinkRow;
-import com.risingwave.proto.Data;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.iceberg.FileFormat;
-import org.apache.iceberg.PartitionSpec;
-import org.apache.iceberg.Schema;
-import org.apache.iceberg.Table;
+import org.apache.iceberg.*;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.data.GenericRecord;
 import org.apache.iceberg.data.IcebergGenerics;
@@ -48,23 +41,14 @@ import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
 import org.junit.Test;
 
-public class IcebergSinkPartitionTest {
+public class IcebergSinkLocalTest {
     static String warehousePath = "/tmp/rw-sinknode/iceberg-sink/warehouse";
     static String databaseName = "demo_db";
-    static String tableName = "demo_table_partitioned";
+    static String tableName = "demo_table";
     static Schema icebergTableSchema =
             new Schema(
                     Types.NestedField.required(1, "id", Types.IntegerType.get()),
-                    Types.NestedField.required(2, "name", Types.StringType.get()),
-                    Types.NestedField.required(3, "part", Types.StringType.get()));
-    static TableSchema tableSchema =
-            new TableSchema(
-                    Lists.newArrayList("id", "name", "part"),
-                    Lists.newArrayList(
-                            Data.DataType.TypeName.INT32,
-                            Data.DataType.TypeName.VARCHAR,
-                            Data.DataType.TypeName.VARCHAR),
-                    Lists.newArrayList("id"));
+                    Types.NestedField.required(2, "name", Types.StringType.get()));
 
     private void createMockTable() throws IOException {
         if (!Paths.get(warehousePath).toFile().isDirectory()) {
@@ -77,7 +61,7 @@ public class IcebergSinkPartitionTest {
         } catch (Exception e) {
             // Ignored.
         }
-        PartitionSpec spec = PartitionSpec.builderFor(icebergTableSchema).identity("part").build();
+        PartitionSpec spec = PartitionSpec.unpartitioned();
         catalog.createTable(tableIdent, icebergTableSchema, spec, Map.of("format-version", "2"));
         catalog.close();
     }
@@ -109,11 +93,9 @@ public class IcebergSinkPartitionTest {
         for (Row row : rows) {
             int id = row.getInt(0);
             String name = row.getString(1);
-            String part = row.getString(2);
             Record record = GenericRecord.create(icebergTableSchema);
             record.setField("id", id);
             record.setField("name", name);
-            record.setField("part", part);
             actual.add(record);
         }
         assertEquals(expected.size(), actual.size());
@@ -128,24 +110,23 @@ public class IcebergSinkPartitionTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        tableSchema,
+                        TableSchema.getMockTableSchema(),
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
 
         try {
-            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 1, "Alice", "aaa")));
+            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 1, "Alice")));
             sink.sync();
 
             Record record1 = GenericRecord.create(icebergTableSchema);
             record1.setField("id", 1);
             record1.setField("name", "Alice");
-            record1.setField("part", "aaa");
             Set<Record> expected = Sets.newHashSet(record1);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
 
-            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 2, "Bob", "bbb")));
+            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 2, "Bob")));
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
 
@@ -154,7 +135,6 @@ public class IcebergSinkPartitionTest {
             Record record2 = GenericRecord.create(icebergTableSchema);
             record2.setField("id", 2);
             record2.setField("name", "Bob");
-            record2.setField("part", "bbb");
             expected.add(record2);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
@@ -173,7 +153,7 @@ public class IcebergSinkPartitionTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        tableSchema,
+                        TableSchema.getMockTableSchema(),
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
@@ -181,18 +161,16 @@ public class IcebergSinkPartitionTest {
         try {
             sink.write(
                     Iterators.forArray(
-                            new ArraySinkRow(Op.INSERT, 1, "Alice", "aaa"),
-                            new ArraySinkRow(Op.INSERT, 2, "Bob", "bbb")));
+                            new ArraySinkRow(Op.INSERT, 1, "Alice"),
+                            new ArraySinkRow(Op.INSERT, 2, "Bob")));
             sink.sync();
 
             Record record1 = GenericRecord.create(icebergTableSchema);
             record1.setField("id", 1);
             record1.setField("name", "Alice");
-            record1.setField("part", "aaa");
             Record record2 = GenericRecord.create(icebergTableSchema);
             record2.setField("id", 2);
             record2.setField("name", "Bob");
-            record2.setField("part", "bbb");
             Set<Record> expected = Sets.newHashSet(record1, record2);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
@@ -211,7 +189,7 @@ public class IcebergSinkPartitionTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        tableSchema,
+                        TableSchema.getMockTableSchema(),
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
