+++ java/connector-node/risingwave-connector-test/src/test/java/com/risingwave/connector/sink/iceberg/IcebergSinkPartitionTest.java
@@ -12,22 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.risingwave.connector;
+package com.risingwave.connector.sink.iceberg;
 
 import static com.risingwave.proto.Data.*;
 import static org.junit.Assert.*;
 import static org.junit.Assert.assertEquals;
 
 import com.google.common.collect.Iterators;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import com.risingwave.connector.IcebergSink;
 import com.risingwave.connector.api.TableSchema;
 import com.risingwave.connector.api.sink.ArraySinkRow;
+import com.risingwave.proto.Data;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.*;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.iceberg.*;
+import org.apache.iceberg.FileFormat;
+import org.apache.iceberg.PartitionSpec;
+import org.apache.iceberg.Schema;
+import org.apache.iceberg.Table;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.data.GenericRecord;
 import org.apache.iceberg.data.IcebergGenerics;
@@ -40,14 +49,23 @@ import org.apache.spark.sql.Row;
 import org.apache.spark.sql.SparkSession;
 import org.junit.Test;
 
-public class IcebergSinkLocalTest {
+public class IcebergSinkPartitionTest {
     static String warehousePath = "/tmp/rw-sinknode/iceberg-sink/warehouse";
     static String databaseName = "demo_db";
-    static String tableName = "demo_table";
+    static String tableName = "demo_table_partitioned";
     static Schema icebergTableSchema =
             new Schema(
                     Types.NestedField.required(1, "id", Types.IntegerType.get()),
-                    Types.NestedField.required(2, "name", Types.StringType.get()));
+                    Types.NestedField.required(2, "name", Types.StringType.get()),
+                    Types.NestedField.required(3, "part", Types.StringType.get()));
+    static TableSchema tableSchema =
+            new TableSchema(
+                    Lists.newArrayList("id", "name", "part"),
+                    Lists.newArrayList(
+                            Data.DataType.TypeName.INT32,
+                            Data.DataType.TypeName.VARCHAR,
+                            Data.DataType.TypeName.VARCHAR),
+                    Lists.newArrayList("id"));
 
     private void createMockTable() throws IOException {
         if (!Paths.get(warehousePath).toFile().isDirectory()) {
@@ -60,7 +78,7 @@ public class IcebergSinkLocalTest {
         } catch (Exception e) {
             // Ignored.
         }
-        PartitionSpec spec = PartitionSpec.unpartitioned();
+        PartitionSpec spec = PartitionSpec.builderFor(icebergTableSchema).identity("part").build();
         catalog.createTable(tableIdent, icebergTableSchema, spec, Map.of("format-version", "2"));
         catalog.close();
     }
@@ -92,9 +110,11 @@ public class IcebergSinkLocalTest {
         for (Row row : rows) {
             int id = row.getInt(0);
             String name = row.getString(1);
+            String part = row.getString(2);
             Record record = GenericRecord.create(icebergTableSchema);
             record.setField("id", id);
             record.setField("name", name);
+            record.setField("part", part);
             actual.add(record);
         }
         assertEquals(expected.size(), actual.size());
@@ -109,23 +129,24 @@ public class IcebergSinkLocalTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        TableSchema.getMockTableSchema(),
+                        tableSchema,
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
 
         try {
-            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 1, "Alice")));
+            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 1, "Alice", "aaa")));
             sink.sync();
 
             Record record1 = GenericRecord.create(icebergTableSchema);
             record1.setField("id", 1);
             record1.setField("name", "Alice");
+            record1.setField("part", "aaa");
             Set<Record> expected = Sets.newHashSet(record1);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
 
-            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 2, "Bob")));
+            sink.write(Iterators.forArray(new ArraySinkRow(Op.INSERT, 2, "Bob", "bbb")));
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
 
@@ -134,6 +155,7 @@ public class IcebergSinkLocalTest {
             Record record2 = GenericRecord.create(icebergTableSchema);
             record2.setField("id", 2);
             record2.setField("name", "Bob");
+            record2.setField("part", "bbb");
             expected.add(record2);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
@@ -152,7 +174,7 @@ public class IcebergSinkLocalTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        TableSchema.getMockTableSchema(),
+                        tableSchema,
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
@@ -160,16 +182,18 @@ public class IcebergSinkLocalTest {
         try {
             sink.write(
                     Iterators.forArray(
-                            new ArraySinkRow(Op.INSERT, 1, "Alice"),
-                            new ArraySinkRow(Op.INSERT, 2, "Bob")));
+                            new ArraySinkRow(Op.INSERT, 1, "Alice", "aaa"),
+                            new ArraySinkRow(Op.INSERT, 2, "Bob", "bbb")));
             sink.sync();
 
             Record record1 = GenericRecord.create(icebergTableSchema);
             record1.setField("id", 1);
             record1.setField("name", "Alice");
+            record1.setField("part", "aaa");
             Record record2 = GenericRecord.create(icebergTableSchema);
             record2.setField("id", 2);
             record2.setField("name", "Bob");
+            record2.setField("part", "bbb");
             Set<Record> expected = Sets.newHashSet(record1, record2);
             validateTableWithIceberg(expected);
             validateTableWithSpark(expected);
@@ -188,7 +212,7 @@ public class IcebergSinkLocalTest {
         TableIdentifier tableIdentifier = TableIdentifier.of(databaseName, tableName);
         IcebergSink sink =
                 new IcebergSink(
-                        TableSchema.getMockTableSchema(),
+                        tableSchema,
                         hadoopCatalog,
                         hadoopCatalog.loadTable(tableIdentifier),
                         FileFormat.PARQUET);
